Just python things
You might be wondering: "How did he not spot this? You are missing a
whole variable!" But no, it works, because when running generator
`config` is defined in the beginning. When calling from a different file
however, this obviously fails. Additionally, my linter/autocompleter does
not catch this because it technically works. "I am not surprised, but
still disappointed"