Preprocessing fix (#19)
[preprocessing] Change preprocessor to account for new compiler errors

LLVM dropped a feature of C89 that automatically declares/creates a
prototype for functions that have not been declared.
Preprocssing did depend on that behaviour as it removed everything
related to external function calls. It did so to ensure that preprocessed
code is compatible across compilers.

We remove some of declarations which have attributes other compilers
don't know about. This works for CSmith code but might break down for 
other projects.