[checker] fix empty body matching
With the recent change of using O0 for the sanitizer, it turns out that
we only allowed cases where clang -O1 was able to eliminate all markers
that had no void as an argument i.e. All DCEMarker() must be removed.

This was because the procedure to add empty bodies to the markers only
matched on marker of the form DCEMarker(void) and not DCEMarker().

Effectively, this made the job of creduce more difficult.